removed tools files and other requested changes
