+++ api/src/index.ts
@@ -9,9 +9,7 @@ export * from './getAzExtResourceType';
 export * from './resources/azure';
 export * from './resources/base';
 export * from './resources/resourcesApi';
-export * from './resources/tenant';
 export * from './resources/workspace';
 export * from './utils/apiUtils';
 export * from './utils/getApi';
 export * from './utils/wrapper';
-
