+++ package.json
@@ -689,7 +689,7 @@
         "@types/request-promise": "^4.1.51",
         "@types/semver": "^7.3.12",
         "@types/uuid": "^9.0.1",
-        "@types/vscode": "^1.81.0",
+        "@types/vscode": "^1.93.0",
         "@types/ws": "^8.5.10",
         "@typescript-eslint/eslint-plugin": "^5.59.11",
         "@vscode/test-electron": "^2.3.8",
@@ -713,7 +713,7 @@
         "@azure/arm-resources": "^5.2.0",
         "@azure/arm-resources-profile-2020-09-01-hybrid": "^2.1.0",
         "@azure/ms-rest-js": "^2.7.0",
-        "@microsoft/vscode-azext-azureauth": "^2.5.0",
+        "@microsoft/vscode-azext-azureauth": "file:../vscode-azuretools/auth/microsoft-vscode-azext-azureauth-2.5.0.tgz",
         "@microsoft/vscode-azext-azureutils": "^2.0.0",
         "@microsoft/vscode-azext-utils": "^2.5.6",
         "buffer": "^6.0.3",
@@ -724,8 +724,5 @@
         "vscode-nls": "^5.0.1",
         "vscode-uri": "^3.0.7",
         "ws": "^8.17.1"
-    },
-    "enabledApiProposals": [
-        "authGetSessions"
-    ]
+    }
 }
