--- src/commands/accounts/AzureSubscription.ts
@@ -1,70 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-import type { TokenCredential } from '@azure/core-auth';
-import type { Environment } from '@azure/ms-rest-azure-env';
-import { AzureAuthentication } from '@microsoft/vscode-azext-azureauth';
-import * as vscode from "vscode";
-
-/**
- * Taken from Azure tools auth provider to make it easier to test changes remove once changes to the tenants view are done
- */
-
-
-
-/**
- * A type representing an Azure subscription ID, not including the tenant ID.
- */
-export type SubscriptionId = string;
-
-/**
- * A type representing an Azure tenant ID.
- */
-export type TenantId = string;
-
-/**
- * Represents an Azure subscription.
- */
-export interface AzureSubscription {
-    /**
-     * Access to the authentication session associated with this subscription.
-     */
-    readonly authentication: AzureAuthentication;
-
-    /**
-     * The Azure environment to which this subscription belongs.
-     */
-    readonly environment: Environment;
-
-    /**
-     * Whether this subscription belongs to a custom cloud.
-     */
-    readonly isCustomCloud: boolean;
-
-    /**
-     * The display name of this subscription.
-     */
-    readonly name: string;
-
-    /**
-     * The ID of this subscription.
-     */
-    readonly subscriptionId: SubscriptionId;
-
-    /**
-     * The ID of the tenant to which this subscription belongs.
-     */
-    readonly tenantId: TenantId;
-
-    /**
-     * The credential for authentication to this subscription. Compatible with Azure track 2 SDKs.
-     */
-    readonly credential: TokenCredential;
-
-    /**
-     * The account associated with this subscription. This is optional as we only need the account if there are duplicate subscriptions.
-     */
-    readonly account?: vscode.AuthenticationSessionAccountInformation;
-}
