--- src/commands/accounts/getSessionFromVSCode.ts
@@ -1,54 +1,0 @@
-/*---------------------------------------------------------------------------------------------
-*  Copyright (c) Microsoft Corporation. All rights reserved.
-*  Licensed under the MIT License. See License.txt in the project root for license information.
-*--------------------------------------------------------------------------------------------*/
-
-import { getConfiguredAuthProviderId, getConfiguredAzureEnv } from "@microsoft/vscode-azext-azureauth";
-import * as vscode from "vscode";
-
-function ensureEndingSlash(value: string): string {
-    return value.endsWith('/') ? value : `${value}/`;
-}
-
-function getResourceScopes(scopes?: string | string[]): string[] {
-    if (scopes === undefined || scopes === "" || scopes.length === 0) {
-        scopes = ensureEndingSlash(getConfiguredAzureEnv().managementEndpointUrl);
-    }
-    const arrScopes = (Array.isArray(scopes) ? scopes : [scopes])
-        .map((scope) => {
-            if (scope.endsWith('.default')) {
-                return scope;
-            } else {
-                return `${scope}.default`;
-            }
-        });
-    return Array.from(new Set<string>(arrScopes));
-}
-
-function addTenantIdScope(scopes: string[], tenantId: string): string[] {
-    const scopeSet = new Set<string>(scopes);
-    scopeSet.add(`VSCODE_TENANT:${tenantId}`);
-    return Array.from(scopeSet);
-}
-
-function getScopes(scopes: string | string[] | undefined, tenantId?: string): string[] {
-    let scopeArr = getResourceScopes(scopes);
-    if (tenantId) {
-        scopeArr = addTenantIdScope(scopeArr, tenantId);
-    }
-    return scopeArr;
-}
-
-/**
- * Wraps {@link vscode.authentication.getSession} and handles:
- * * Passing the configured auth provider id
- * * Getting the list of scopes, adding the tenant id to the scope list if needed
- *
- * @param scopes - top-level resource scopes (e.g. http://management.azure.com, http://storage.azure.com) or .default scopes. All resources/scopes will be normalized to the `.default` scope for each resource.
- * @param tenantId - (Optional) The tenant ID, will be added to the scopes
- * @param options - see {@link vscode.AuthenticationGetSessionOptions}
- * @returns An authentication session if available, or undefined if there are no sessions
- */
-export async function getSessionFromVSCode(scopes?: string | string[], tenantId?: string, options?: vscode.AuthenticationGetSessionOptions): Promise<vscode.AuthenticationSession | undefined> {
-    return await vscode.authentication.getSession(getConfiguredAuthProviderId(), getScopes(scopes, tenantId), options);
-}
