+++ src/tree/ResourceGroupsTreeContext.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { ISubscriptionContext } from "@microsoft/vscode-azext-utils";
-import { AzureSubscription } from "../commands/accounts/AzureSubscription"; //changes within this file will eventually be in the auth package
+import { AzureSubscription } from "../../api/src/index";
 import { ResourceGroupsItem } from "./ResourceGroupsItem";
 
 export interface ResourceGroupsTreeContext {
