+++ src/tree/tenants/tenant.ts
@@ -3,7 +3,7 @@
 *  Licensed under the MIT License. See License.md in the project root for license information.
 *--------------------------------------------------------------------------------------------*/
 
-import { BranchDataProvider, ResourceBase, ResourceModelBase, ResourceProvider } from "./base";
+import { BranchDataProvider, ResourceBase, ResourceModelBase, ResourceProvider } from "api/src";
 
 /**
  * Respresents a specific type of tenant resource.
