+++ src/utils/getResourceContributions.ts
@@ -15,10 +15,6 @@ interface ResourceGroupsContribution {
         readonly branches?: { type: string }[];
         readonly resources?: boolean;
     }
-    readonly tenant: {
-        readonly branches?: { type: string }[];
-        readonly resources?: boolean;
-    }
     readonly commands?: (vscode.Command & { detail?: string })[];
 }
